Move group state module into group module and rename
