+++ p2panda-auth/src/group/test_utils/network.rs
@@ -6,8 +6,7 @@ use rand::SeedableRng;
 use rand::rngs::StdRng;
 use rand::seq::SliceRandom;
 
-use crate::group::{GroupAction, GroupControlMessage, GroupMember};
-use crate::group_crdt::Access;
+use crate::group::{GroupAction, GroupControlMessage, GroupMember, Access};
 use crate::traits::{AuthGraph, GroupStore, Operation, Ordering};
 
 use super::{
